+++ Library/OcMainLib/OpenCoreMisc.c
@@ -977,8 +977,6 @@ OcMiscBoot (
 
   OcLoadPickerHotKeys (Context);
 
-  Context->ShowToggleSip   = Config->Misc.Security.AllowToggleSip;
-  Context->ShowNvramReset  = Config->Misc.Security.AllowNvramReset;
   Context->AllowSetDefault = Config->Misc.Security.AllowSetDefault;
   if (!Config->Misc.Security.AllowNvramReset && (Context->PickerCommand == OcPickerResetNvram)) {
     Context->PickerCommand = PickerCommand;